Extract ZMQ adapter from JustMocks
